pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["Gabriel Carvajal <36053537+gcarvajal1222@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
